+++ srcs/utils/case_proof.cpp
@@ -6,7 +6,7 @@
 /*   By: abesombe <abesombe@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/13 15:20:32 by abesombe          #+#    #+#             */
-/*   Updated: 2022/04/14 12:00:46 by abesombe         ###   ########.fr       */
+/*   Updated: 2022/04/18 18:33:41 by abesombe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 std::string case_proof(std::string str)
 {
-    int size = str.length();
+    size_t size = str.length();
     char char_str[size + 1]; 
-    for (int i = 0; i < str.length(); i++)
+    for (size_t i = 0; i < str.length(); i++)
     {
         if (str[i] >= 'a' && str[i] <= '~')
             char_str[i] = str.c_str()[i] - 32;
